Fix auto-completion in several drop-down fields
In all these cases the matchProp needs to be set to "label" and not "value"
because the value contains the object ID or another identifier.

After this, the auto-completion is looking at same data that the user
is seeing.

Refs Graylog2/graylog2-server#5659

(cherry picked from commit f30d17e343b5a401a2360bcf6f17931c96ea211e)